fix all problems

+++ services/fighterService.js
@@ -1,14 +1,6 @@
 import { fighterRepository } from "../repositories/fighterRepository.js";
 
 class FighterService {
-  getAll() {
-    const fighters = fighterRepository.getAll();
-    if (!fighters) {
-      return null;
-    }
-    return fighters;
-  }
-
   search(search) {
     const item = fighterRepository.getOne(search);
     if (!item) {
@@ -17,16 +9,32 @@ class FighterService {
     return item;
   }
 
+  getAll() {
+    const fighters = fighterRepository.getAll();
+    if (!fighters) {
+      throw new Error("Can't find fighters!");
+    }
+    return fighters;
+  }
+
+  getOne(id) {
+    const fighter = this.search({ id: id });
+    if (!fighter) {
+      throw new Error("Can't find fighter!");
+    }
+    return fighter;
+  }
+
   create(data) {
     const { name, health } = data;
     const fighterExists = this.search({ name: name });
     if (fighterExists) {
-      return null;
+      throw new Error(`Fighter ${name} exists!`);
     }
     data.health = health ?? 100;
     const newFighter = fighterRepository.create(data);
     if (!newFighter) {
-      return null;
+      throw new Error("Can't create fighter!");
     }
     return newFighter;
   }
@@ -35,23 +43,27 @@ class FighterService {
     const { name } = data;
     const fighterExists = this.search({ name: name });
     if (fighterExists) {
-      return null;
+      throw new Error(`Fighter ${name} exists!`);
     }
     const fighterForUpdate = this.search({ id: id });
     if (!fighterForUpdate) {
-      return null;
+      throw new Error("Can't find fighter!");
     }
     const updateFighter = fighterRepository.update(id, data);
     if (!updateFighter) {
-      return null;
+      throw new Error("Can't find fighter!");
     }
     return updateFighter;
   }
 
   delete(id) {
+    const fighterForDelete = this.search({ id: id });
+    if (!fighterForDelete) {
+      throw new Error("Can't find fighter!");
+    }
     const deleteFighter = fighterRepository.delete(id);
     if (!deleteFighter) {
-      return null;
+      throw new Error("Can't find fighter!");
     }
     return deleteFighter;
   }
